+++ changelog.md
@@ -6,11 +6,13 @@
   * Minimum hunger for players to sprint, has been reduced to 4 from vanilla 7
 * Tiredness.Tired Overlay
   * Enables/Disables the overlay with the tired effect
+* Config options for raw food poison amplifier and duration
 * Timber Trees support for Environmental and Atmospheric
 
 ### Changes
 * Sandstone recipe now gives 2 instead of 1
 * Bundle recipe is now vanilla one but with cloth (1 string + 1 cloth)
+* Increased raw food poison effectiveness
 
 ### Removed
 * Shields+ shield definitions in Shields+ Compat Data pack